Tweaked openssl.mk to take an optional make argument OPENSSL_VERSION. If OPENSSL_VERSION is not specified, then openssl.mk does exactly what it did previously, e.g., makes openssl 0.9.7 or 0.9.8 depending on target. If OPENSSL_VERSION is specified, like this:
   make OPENSSL_VERSION=1.0.1 openssl[-stage | -ipk]

then openssl.mk will make that version of openssl. 

Note that if you use the ipk file to install openssl 1.0.x, it will remove any older version of libcrypto and 
libssl. So, copy libcrypto.so.0.9.x and libssl.0.9.x to a safe place, install openssl 1.0.x, copy libcrypto.so.0.9.x
and libssl.so.0.9.x back again and manually fix up the symbolic links.

Also note that the openssl 1.0.1 Makefile names the libraries 1.0.0.


git-svn-id: http://svn.nslu2-linux.org/svnroot/optware/trunk@12776 1f1dca69-5516-0410-b834-e485effd04c3